- "number" is now translated to "int". - assert base_url, public_key, and private_key are not None before using them.

+++ src/ops_manager_sdk/generator/utils.py
@@ -122,7 +122,7 @@ def type_mapping(type_str: str) -> str:
         "string": "str",
         "integer": "int",
         "long": "int",
-        "number": "float",
+        "number": "int",
         "boolean": "bool",
         "object": "dict",
         "timestamp": "str",
